respaldo 2/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
PyJWT==2.8.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
openpyxl==3.1.2
beautifulsoup4==4.12.3
requests==2.31.0
Flask-Limiter==3.5.0
